+++ engine/src/main/java/net/jqwik/engine/facades/ArbitraryFacadeImpl.java
@@ -125,20 +125,7 @@ public class ArbitraryFacadeImpl extends Arbitrary.ArbitraryFacade {
 
 	@Override
 	public <T> Stream<T> sampleStream(Arbitrary<T> arbitrary) {
-		RandomGenerator<T> generator = getGeneratorForSampling(arbitrary);
-		return generator
-					   .stream(SourceOfRandomness.current())
-					   .map(Shrinkable::value);
-	}
-
-	private static final Map<Arbitrary<Object>, RandomGenerator<Object>> generators = new HashMap<>();
-
-	@SuppressWarnings("unchecked")
-	private static <T> RandomGenerator<T> getGeneratorForSampling(Arbitrary<T> arbitrary) {
-		return (RandomGenerator<T>) generators.computeIfAbsent(
-				(Arbitrary<Object>) arbitrary,
-				a -> a.generator(JqwikProperties.DEFAULT_TRIES, true)
-		);
+		return new SampleStreamFacade().sampleStream(arbitrary);
 	}
 
 	@Override